fix[okta-vue]: Fixes login_required error (#304)
The TokenManager throws an error when tries to renew a token but Okta session is expired.
The SDK should capture that error in the getAccessToken and getIdToken functions and return undefined instead.
In this way, also the isAuthenticated function will return false, so the router can correctly redirect to a new login.

Also makes the isAuthenticated function to use the getAccessToken() and getIdToken() functions instead of using directly the tokenManager, like we do for the other SDKs.